--- types/index.js
@@ -1,1 +1,0 @@
-// Intentionally empty.
